Refine documentation of segmentation model testing
Updated the documentation of a recent experiment with segmentation models. The commit provides a more concise explanation of the issues encountered during the test of Deeplabv3 and yolov8-seg models, the modifications made to the code for better results, as well as the process of model training and validation set prediction. The aim is to maintain a clear and concise record of the methodology and improve readability for future referencing.